alpha/requirements.txt
selenium
python-binance